common auth service
can easily be extended if more auth options are available
